requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
django==2.2.5
djangorestframework==3.10.3
pytz==2019.2              # via django
sqlparse==0.3.0           # via django
